pyproject.toml
[project]
name = "evals"
version = "0.1.0"
description = "Copilot Studio Eval Platform — manage datasets, run evals, view results"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "reflex>=0.8.27",
    "microsoft-agents-copilotstudio-client>=0.7.0",
    "msal>=1.31.0",
    "deepeval>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
